feat: Modernize application and add new features
This commit modernizes the job portal application by restructuring the codebase, enhancing user authentication, improving job listings and search, and adding a resume upload feature.

- **Modernized Application Architecture:** The application is restructured into a more modular format with separate files for different concerns.
- **Enhanced User Authentication and Management:** A user profile page has been added, allowing you to view and update your information and see a list of jobs you have applied for.
- **Improved Job Listings and Search:** The job listings page now includes more advanced search and filtering options, as well as pagination.
- **Added Resume/CV Upload Feature:** You can now upload your resumes, which are stored in the database using GridFS.
- **Implemented Robust Notification System:** A more robust notification system using Celery has been implemented to send emails asynchronously.
- **Updated Dependencies and Environment Management:** The `requirements.txt` file has been updated, and a `README.md` file with setup instructions has been created.
- **Refined UI and User Experience:** The UI has been refined with the addition of icons to the sidebar navigation.
