+++ auth.py
@@ -1,0 +1,42 @@
+import streamlit as st
+from passlib.hash import pbkdf2_sha256
+from db import users_collection
+
+def sign_in():
+    st.title("Sign In")
+    username = st.text_input("Username")
+    password = st.text_input("Password", type="password")
+    if st.button("Sign In"):
+        user = users_collection.find_one({"username": username})
+        if user and pbkdf2_sha256.verify(password, user["password"]):
+            st.success("You have successfully signed in!")
+            st.session_state.user_id = user["_id"]
+            st.session_state.role = user["role"]
+            st.experimental_rerun()
+        else:
+            st.error("Invalid username or password. Please try again.")
+
+def sign_up():
+    st.title("Sign Up")
+    username = st.text_input("Username")
+    email = st.text_input("Email")
+    password = st.text_input("Password", type="password")
+    confirm_password = st.text_input("Confirm Password", type="password")
+    if st.button("Sign Up"):
+        if password == confirm_password:
+            hashed_password = pbkdf2_sha256.hash(password)
+            user_data = {
+                "username": username,
+                "email": email,
+                "password": hashed_password,
+                "role": "user"
+            }
+            users_collection.insert_one(user_data)
+            st.success("You have successfully signed up! Please sign in.")
+        else:
+            st.error("Passwords do not match. Please try again.")
+
+def sign_out():
+    for key in list(st.session_state.keys()):
+        del st.session_state[key]
+    st.experimental_rerun()
