+++ requirements.txt
@@ -2,5 +2,7 @@ streamlit
 pymongo
 passlib
 python-dotenv
+celery
+redis
